Add files via upload
One of the top user stories on my board is encouraging feedback for a high score, so the user would be inclined to play again.  Inspirational quotes seemed to be a neat way to go, since even a perfect score would have a random payoff. A future variation on the same idea might be combining libraries of different clauses into a single grammatical sentence, which might be less copyright infringe-y.